refactor: move format json function to another ts file
