+++ frontend/src/lib/multipleDeleteFormating.ts
@@ -1,0 +1,21 @@
+interface MultipleDeleteSubscriberData {
+  ueId: string;
+  plmnID: string;
+}
+
+export function formatMultipleDeleteSubscriberToJson(subscribers: MultipleDeleteSubscriberData[]) {
+  return subscribers.map(sub => ({
+    ueId: sub.ueId,
+    plmnID: sub.plmnID
+  }));
+}
+
+interface MultipleDeleteProfileData {
+  profileName: string;
+}
+
+export function formatMultipleDeleteProfileToJson(profiles: MultipleDeleteProfileData[]) {
+  return profiles.map(profile => ({
+    profileName: profile.profileName
+  }));
+}
